MVP-5815-1: extrct useTargetListItem hook logic
